requirements.txt
Django==1.9.5
psycopg2==2.6.1
django-extensions==1.6.3
django-bower==5.1.0
django-filter==0.13.0
django-model-utils==2.5
djangorestframework==3.3.3
lxml==3.6.0
Markdown==2.6.6
requests==2.10.0